+++ src/extension.ts
@@ -5,22 +5,42 @@ class Settings {
 	getURL() {
 		return vscode.workspace.getConfiguration("devdocs-io").get("URL", "https://devdocs.io/").toString();
 	}
+
+	getСolumn() {
+		let column = vscode.workspace.getConfiguration("devdocs-io").get("column", "Beside").toString();
+		switch(column) {
+			case "Active":
+				return vscode.ViewColumn.Active;
+			case "Beside":
+				return vscode.ViewColumn.Beside;
+			case "First":
+				return vscode.ViewColumn.One;
+			case "Second":
+				return vscode.ViewColumn.Two;
+			case "Third":
+				return vscode.ViewColumn.Three;
+			default:
+				return vscode.ViewColumn.Beside;
+		}
+	}
 }
 
 class Panel {
 	id: number;
 	removed: boolean;
+	settings: Settings;
 	parent: PanelStorage;
 	panel: vscode.WebviewPanel;
 
-    constructor(parent: PanelStorage) {
+    constructor(settings: Settings, parent: PanelStorage) {
 		this.id = 1;
 		this.removed = false;
+		this.settings = settings;
 		this.parent = parent;
 		this.panel = vscode.window.createWebviewPanel(
-			'devdocs',
-			'devdocs',
-			vscode.ViewColumn.Two,
+			'devdocs.io',
+			'devdocs.io',
+			this.settings.getСolumn(),
 			{
 				enableScripts: true,
 				retainContextWhenHidden: true
@@ -53,16 +73,18 @@ class Panel {
 
 		if (!this.removed) {
 			this.panel.webview.html = html;
-			this.panel.reveal(this.panel.viewColumn);
+			this.panel.reveal(this.settings.getСolumn());
 		}
 	}
 }
 
 class PanelStorage {
 	panels: Panel[];
+	settings: Settings;
 
-	constructor() {
+	constructor(settings: Settings) {
 		this.panels = [];
+		this.settings = settings;
 	}
 
 	onDestroyPanel() {
@@ -74,7 +96,7 @@ class PanelStorage {
 	}
 
 	add(url: string) {
-		let panel = new Panel(this);
+		let panel = new Panel(this.settings, this);
 		panel.goto(url);
 		this.panels.push(panel);
 	}
@@ -90,8 +112,8 @@ class PanelStorage {
 }
 
 export function activate(context: vscode.ExtensionContext) {
-	let panels = new PanelStorage();
 	let settings = new Settings();
+	let panels = new PanelStorage(settings);
 
 	let searchCmd = vscode.commands.registerCommand('devdocs-io.search', () => {
 		panels.last(settings.getURL());