+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.1.0
 
 - Added panel with devdocs.io
-- Added settings: "URL"
+- Added settings: "url" and "column"
 
 ## 0.0.1